requirements.txt
Django>=5.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
